Moved `user` helper to `helpers`
no refs

Moved the user helper to helpers and updated all references to it
