+++ src/handlers.ts
@@ -18,7 +18,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { addToList, removeFromList } from './kv-helpers';
 import { ContextData, HonoContextVariables, fedify } from './app';
 import { getSiteSettings } from './helpers/ghost';
-import type { PersonData } from './user';
+import type { PersonData } from './helpers/user';
 import { ACTOR_DEFAULT_HANDLE } from './constants';
 import { Temporal } from '@js-temporal/polyfill';
 import { createHash } from 'node:crypto';
